Backend/requirements.txt
fastapi
uvicorn
pyyaml
langchain
langchain-huggingface
python-dotenv
